Move Messages to Abstractions
#44 breaking
