+++ src/Violet.WorkItems.Abstractions/ValidationModel.cs
@@ -1,0 +1,3 @@
+namespace Violet.WorkItems.Validation;
+
+public record ErrorMessage(string Source, string ErrorCode, string Message, string ProjectCode, string Id, string Property);
