+++ src/Violet.WorkItems.Abstractions/ValueProviderModel.cs
@@ -1,0 +1,5 @@
+using System;
+
+namespace Violet.WorkItems.ValueProvider;
+
+public record ProvidedValue(string Value, string DisplayText, bool IsAllowed = true);
