+++ src/Violet.WorkItems.BlazorWebFrontend/Pages/WorkItemPage.razor
@@ -3,6 +3,7 @@
 @inject WorkItemService WorkItemService
 @inject WorkItemDescriptorService WorkItemDescriptorService
 @inject NavigationManager NavigationManager
+@using Violet.WorkItems.Service.Messages;
 @using Violet.WorkItems.Types
 @using Violet.WorkItems.BlazorWebFrontend.Shared.Widgets
 @using Violet.WorkItems.BlazorWebFrontend.Shared.RazorBlades
