+++ src/Violet.WorkItems.BlazorWebFrontend/Shared/Widgets/PropertyWidget.razor
@@ -1,5 +1,6 @@
-@using Violet.WorkItems.Types
 @using System.Linq
+@using Violet.WorkItems.Types
+@using Violet.WorkItems.Service.Messages;
 
 <DynamicComponent Type="@GetWidget(PropertyDescriptor)"
     Parameters="GetWidgetParameters(PropertyDescriptor, WorkItem)" />
