+++ src/Violet.WorkItems.Service/Controllers/WorkItemController.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
-using Violet.WorkItems.Service.Models;
+using Violet.WorkItems.Service.Messages;
 
 namespace Violet.WorkItems.Service.Controllers;
